markov update solutions and spec
