+++ Labs/Vol2A/MarkovChains/spec.py
@@ -1,63 +1,74 @@
-# spec.py
-"""Volume II Lab 8: Markov Chains
+# name this file 'solutions.py'
+"""Volume II: Markov Chains.
 <Name>
 <Class>
 <Date>
 """
 
+import numpy as np
 
-# Problem 1: implement this function.
+
+# Problem 1
 def random_markov(n):
-    """Create and return a transition matrix for a random
-    Markov chain with 'n' states as an nxn NumPy array.
+    """Create and return a transition matrix for a random Markov chain with
+    'n' states. This should be stored as an nxn NumPy array.
     """
-    raise NotImplementedError("Problem 1 incomplete.")
+    raise NotImplementedError("Problem 1 Incomplete")
 
 
-# Problem 2: modify this function.
-def forecast(num_days):
-    """Run a simulation for the weather over 'num_days' days, with
-    "hot" as the starting state. Return a list containing the day-by-day
-    results, not including the starting day.
+# Problem 2
+def forecast():
+    """Forecast tomorrow's weather given that today is hot."""
+    
+    transition_matrix = np.array([[.7, .6], [.3, .4]])
+    # Sample from the standard uniform distribution to choose a new state.
+    if np.random.random() < transition_matrix[1, 0]:
+        print("Cold")
+        return 1
+    else:
+        print "Hot"
+        return 0
+
+
+# Problem 3
+def four_state_forecast(days):
+    """Run a simulation for the weather over the specified number of days,
+    with 'mild' as the starting state, using the four-state Markov chain.
+    Return a list containing the day-by-day results, not including the
+    starting day.
+
+    Examples:
+        >>> four_state_forecast(3)
+        [0, 1, 3]
+        >>> four_state_forecast(5)
+        [2, 1, 2, 1, 1]
+    """
+    raise NotImplementedError("Problem 3 Incomplete")
+
+
+# Problem 4
+def analyze_simulation():
+    """Investigate and interpret the results of the simulations in the previous
+    two problems. Specifically, find the average percentage of days that are
+    hot, mild, cold, and freezing in each simulation. Does the starting day
+    alter the results? Print a report of your findings (return nothing).
+    """
+    raise NotImplementedError("Problem 4 Incomplete")
+
+
+# Problems 5 and 6
+class SentenceGenerator(object):
+    """Note: This implementation is ROW STOCHASTIC!!!
 
     Example:
-        >>> forecast(5)
-        [1, 0, 0, 1, 0]
-        # Or, if you prefer,
-        ['cold', 'hot', 'hot', 'cold', 'hot']
+        >>> yoda = SentenceGenerator("Yoda.txt")
+        >>> print yoda.babble()
+        The dark side of loss is a path as one with you.
     """
-    raise NotImplementedError("Problem 2 incomplete.")
 
+    def __init__(self, filename):
+        raise NotImplementedError("Problem 5 Incomplete")
 
-# Problem 3: implement this function.
-def four_state_forecast(days=1):
-    """Same as forecast(), but using the four-state transition matrix."""
-    raise NotImplementedError("Problem 3 incomplete.")
+    def babble(self):
+        raise NotImplementedError("Problem 6 Incomplete")
 
-
-# Problem 4: implement this function.
-def analyze_simulation():
-    """Analyze the results of the previous two problems. What percentage
-    of days are in each state? Print your results to the terminal.
-    """
-    raise NotImplementedError("Problem 4 incomplete.")
-
-
-# Problems 5-6: define and implement the described functions.
-
-
-# Problem 7: implement this function.
-def sentences(infile, outfile, num_sentences=1):
-    """Generate random sentences using the word list generated in
-    Problem 5 and the transition matrix generated in Problem 6.
-    Write the results to the specified outfile.
-
-    Parameters:
-        infile (str): The path to a filen containing a training set.
-        outfile (str): The file to write the random sentences to.
-        num_sentences (int): The number of random sentences to write.
-
-    Returns:
-        None
-    """
-    raise NotImplementedError("Problem 7 incomplete.")
